[#183] create main driver for feature extraction
