+++ src/ml_service/feature_extraction/feature_extraction_test.py
@@ -1,10 +1,10 @@
 import unittest
-from feature_extraction import main
+from feature_extraction import feature_extraction
 
 
 class TestStringMethods(unittest.TestCase):
 
     def test_process_command(self):
         with self.assertRaises(SystemExit) as cm:
-            main.process_command('random', [])
+            feature_extraction.process_command('random', [])
         self.assertEqual(cm.exception.code, 1)
