+++ src/ml_service/feature_extraction/pre_processing/word_vector/french_related_word_fetcher.py
@@ -1,4 +1,3 @@
-import logging
 import os
 import pickle
 import re
